+++ packages/lib/uni-data-picker/uni-data-picker.vue
@@ -10,7 +10,8 @@
 					<scroll-view v-else-if="inputSelected.length" class="selected-area" scroll-x="true">
 						<view class="selected-list">
 							<view class="selected-item" v-for="(item,index) in inputSelected" :key="index">
-								<text>{{item.text}}</text><text v-if="index<inputSelected.length-1" class="input-split-line">{{split}}</text>
+								<text>{{item.text}}</text><text v-if="index<inputSelected.length-1"
+									class="input-split-line">{{split}}</text>
 							</view>
 						</view>
 					</scroll-view>
@@ -32,9 +33,10 @@
 					<view class="dialog-close-plus dialog-close-rotate" data-id="close"></view>
 				</view>
 			</view>
-			<data-picker-view class="picker-view" ref="pickerView" v-model="value" :localdata="localdata" :preload="preload"
-			 :collection="collection" :field="field" :orderby="orderby" :where="where" :step-searh="stepSearh" :self-field="selfField"
-			 :parent-field="parentField" :managed-mode="true" @change="onchange" @datachange="ondatachange" @nodeclick="onnodeclick"></data-picker-view>
+			<data-picker-view class="picker-view" ref="pickerView" v-model="dataValue" :localdata="localdata"
+				:preload="preload" :collection="collection" :field="field" :orderby="orderby" :where="where"
+				:step-searh="stepSearh" :self-field="selfField" :parent-field="parentField" :managed-mode="true"
+				@change="onchange" @datachange="ondatachange" @nodeclick="onnodeclick"></data-picker-view>
 		</view>
 	</view>
 </template>
@@ -68,6 +70,7 @@
 	 */
 	export default {
 		name: 'UniDataPicker',
+		emits: ['popupopened', 'popupclosed', 'nodeclick', 'input', 'change','update:modelValue'],
 		mixins: [dataPicker],
 		components: {
 			DataPickerView
@@ -132,18 +135,18 @@
 			},
 			load() {
 				if (this.readonly) {
-					this._processReadonly(this.localdata, this.value)
+					this._processReadonly(this.localdata, this.dataValue)
 					return
 				}
 
 				if (this.isLocaldata) {
 					this.loadData()
 					this.inputSelected = this.selected.slice(0)
-				} else if (!this.parentField && !this.selfField && this.value) {
+				} else if (!this.parentField && !this.selfField && this.dataValue) {
 					this.getNodeData(() => {
 						this.inputSelected = this.selected.slice(0)
 					})
-				} else if (this.value.length) {
+				} else if (this.dataValue.length) {
 					this.getTreePath(() => {
 						this.inputSelected = this.selected.slice(0)
 					})
@@ -242,13 +245,14 @@
 					value[i] = selected[i].value
 				}
 
-        const item = selected[selected.length - 1]
+				const item = selected[selected.length - 1]
 
 				if (this.formItem) {
 					this.formItem.setValue(item.value)
 				}
 
-        this.$emit('input', item.value)
+				this.$emit('input', item.value)
+				this.$emit('update:modelValue', item.value)
 				this.$emit('change', {
 					detail: {
 						value: selected