+++ packages/lib/uni-file-picker/uni-file-picker.vue
@@ -40,7 +40,7 @@
 </template>
 
 <script>
-
+	import {chooseAndUploadFile} from './choose-and-upload-file.js'
 	import uploadImage from './upload-image.vue'
 	import uploadFile from './upload-file.vue'
 	let fileInput = null
@@ -246,7 +246,11 @@ export default {
 		}
 	},
 	created() {
-		// this.files = Object.assign([], this.value)
+		// TODO 兼容不开通服务空间的情况
+		if(!(uniCloud.config && uniCloud.config.provider)){
+			this.noSpace = true
+			uniCloud.chooseAndUploadFile = chooseAndUploadFile
+		}
 		this.tempData = {}
 		this.form = this.getForm('uniForms')
 		this.formItem = this.getForm('uniFormsItem')
@@ -297,6 +301,7 @@ export default {
 		 * 选择文件
 		 */
 		choose() {
+			
 			if (this.disabled) return
 			if (this.files.length >= Number(this.limitLength) && this.showType !== 'grid' && this.returnType === 'array') {
 				uni.showToast({
@@ -361,6 +366,7 @@ export default {
 							if (this.limitLength - this.files.length <= 0) break
 							files[i].uuid = Date.now()
 							let filedata = await this.getFileData(files[i], this.fileMediatype)
+							filedata.file = files[i]
 							filedata.progress = 0
 							filedata.status = 'ready'
 							this.files.push(filedata)
@@ -372,7 +378,7 @@ export default {
 						})
 						res.tempFiles = files
 						// 停止自动上传
-						if (!this.autoUpload) {
+						if (!this.autoUpload || this.noSpace) {
 							res.tempFiles = []
 							// TODO 先放弃这个实现 ，不能全部上传
 							// return new Promise((resolve) => {
@@ -542,6 +548,7 @@ export default {
 				uuid: files.uuid,
 				extname: extname || '',
 				cloudPath: files.cloudPath,
+				file:files.file,
 				fileType: files.fileType,
 				url: files.path || files.path,
 				size: files.size, //单位是字节
@@ -551,9 +558,12 @@ export default {
 			}
 			if (type === 'image') {
 				const imageinfo = await this.getFileInfo(files.path)
+				delete filedata.video
 				filedata.image.width = imageinfo.width
 				filedata.image.height = imageinfo.height
 				filedata.image.location = imageinfo.path
+			}else{
+				delete filedata.image
 			}
 			return filedata
 		},