+++ packages/lib/uni-popup/uni-popup.vue
@@ -161,10 +161,10 @@ export default {
 		}
 		fixSize()
 		// #ifdef H5
-		window.addEventListener('resize', fixSize)
-		this.$once('hook:beforeDestroy', () => {
-			window.removeEventListener('resize', fixSize)
-		})
+		// window.addEventListener('resize', fixSize)
+		// this.$once('hook:beforeDestroy', () => {
+		// 	window.removeEventListener('resize', fixSize)
+		// })
 		// #endif
 	},
 	created() {