+++ packages/lib/uni-th/filter-dropdown.vue
@@ -1,14 +1,19 @@
 <template>
 	<view class="uni-filter-dropdown">
 		<view class="dropdown-btn" @click="onDropdown">
-			<view class="icon-select" :class="{active: canReset}" v-if="filterType == 'select'"></view>
-			<view class="icon-search" :class="{active: canReset}" v-if="filterType == 'search'">
+			<view class="icon-select" :class="{active: canReset}" v-if="isSelect || isRange"></view>
+			<view class="icon-search" :class="{active: canReset}" v-if="isSearch">
 				<view class="icon-search-0"></view>
 				<view class="icon-search-1"></view>
+			</view>
+			<view class="icon-calendar" :class="{active: canReset}" v-if="isDate">
+				<view class="icon-calendar-0"></view>
+				<view class="icon-calendar-1"></view>
 			</view>
 		</view>
 		<view class="uni-dropdown-cover" v-if="isOpened" @click="handleClose"></view>
 		<view class="dropdown-popup dropdown-popup-right" v-if="isOpened" @click.stop>
+			<!-- select-->
 			<view v-if="isSelect" class="list">
 				<label class="flex-r a-i-c list-item" v-for="(item,index) in dataList" :key="index"
 					@click="onItemClick($event, index)">
@@ -19,16 +24,36 @@
 				</label>
 			</view>
 			<view v-if="isSelect" class="flex-r opera-area">
-				<view class="flex-f btn btn-default" :class="{disable: !canReset}" @click="handleSelectReset">{{resource.reset}}</view>
+				<view class="flex-f btn btn-default" :class="{disable: !canReset}" @click="handleSelectReset">
+					{{resource.reset}}</view>
 				<view class="flex-f btn btn-submit" @click="handleSelectSubmit">{{resource.submit}}</view>
 			</view>
+			<!-- search -->
 			<view v-if="isSearch" class="search-area">
 				<input class="search-input" v-model="filterValue" />
 			</view>
 			<view v-if="isSearch" class="flex-r opera-area">
 				<view class="flex-f btn btn-submit" @click="handleSearchSubmit">{{resource.search}}</view>
-				<view class="flex-f btn btn-default":class="{disable: !canReset}" @click="handleSearchReset">{{resource.reset}}
-				</view>
+				<view class="flex-f btn btn-default" :class="{disable: !canReset}" @click="handleSearchReset">
+					{{resource.reset}}</view>
+			</view>
+			<!-- range -->
+			<view v-if="isRange">
+				<view class="input-label">{{resource.gt}}</view>
+				<input class="input" v-model="gtValue" />
+				<view class="input-label">{{resource.lt}}</view>
+				<input class="input" v-model="ltValue" />
+			</view>
+			<view v-if="isRange" class="flex-r opera-area">
+				<view class="flex-f btn btn-default" :class="{disable: !canReset}" @click="handleRangeReset">
+					{{resource.reset}}</view>
+				<view class="flex-f btn btn-submit" @click="handleRangeSubmit">{{resource.submit}}</view>
+			</view>
+			<!-- date -->
+			<view v-if="isDate">
+				<uni-datetime-picker ref="datetimepicker" :value="dateRange" type="datetimerange" return-type="timestamp" @change="datetimechange" @maskClick="timepickerclose">
+					<view></view>
+				</uni-datetime-picker>
 			</view>
 		</view>
 	</view>
@@ -41,12 +66,18 @@
 		"reset": "重置",
 		"search": "搜索",
 		"submit": "确定",
-		"filter": "筛选"
+		"filter": "筛选",
+		"gt": "大于等于",
+		"lt": "小于等于",
+		"date": "日期范围"
 	}
 
 	const DropdownType = {
 		Select: "select",
-		Search: "search"
+		Search: "search",
+		Range: "range",
+		Date: "date",
+		Timestamp: "timestamp"
 	}
 
 	export default {
@@ -83,12 +114,6 @@
 			}
 		},
 		computed: {
-			isSelect() {
-				return this.filterType === DropdownType.Select
-			},
-			isSearch() {
-				return this.filterType === DropdownType.Search
-			},
 			canReset() {
 				if (this.isSearch) {
 					return this.filterValue.length > 0
@@ -96,7 +121,25 @@
 				if (this.isSelect) {
 					return this.checkedValues.length > 0
 				}
+				if (this.isRange) {
+					return (this.gtValue.length > 0 && this.ltValue.length > 0)
+				}
+				if (this.isDate) {
+					return this.dateSelect.length > 0
+				}
 				return false
+			},
+			isSelect() {
+				return this.filterType === DropdownType.Select
+			},
+			isSearch() {
+				return this.filterType === DropdownType.Search
+			},
+			isRange() {
+				return this.filterType === DropdownType.Range
+			},
+			isDate() {
+				return (this.filterType === DropdownType.Date || this.filterType === DropdownType.Timestamp)
 			}
 		},
 		watch: {
@@ -114,7 +157,11 @@
 				isOpened: false,
 				dataList: [],
 				filterValue: '',
-				checkedValues: []
+				checkedValues: [],
+				gtValue: '',
+				ltValue: '',
+				dateRange: [],
+				dateSelect: []
 			};
 		},
 		created() {
@@ -132,12 +179,25 @@
 			},
 			openPopup() {
 				this.isOpened = true
+				if (this.isDate) {
+					this.$nextTick(() => {
+						if (!this.dateRange.length) {
+							this.resetDate()
+						}
+						this.$refs.datetimepicker.show()
+					})
+				}
 			},
 			closePopup() {
 				this.isOpened = false
 			},
 			handleClose(e) {
 				this.closePopup()
+			},
+			resetDate() {
+				let date = new Date()
+				let dateText = date.toISOString().split('T')[0]
+				this.dateRange = [dateText + ' 0:00:00', dateText + ' 23:59:59']
 			},
 			onDropdown(e) {
 				this.openPopup()
@@ -160,9 +220,22 @@
 				}
 				this.checkedValues = checkvalues
 			},
+			datetimechange(e) {
+				this.closePopup()
+				this.dateRange = e
+				this.dateSelect = e
+				this.$emit('change', {
+					filterType: this.filterType,
+					filter: e
+				})
+			},
+			timepickerclose(e) {
+				this.closePopup()
+			},
 			handleSelectSubmit() {
 				this.closePopup()
 				this.$emit('change', {
+					filterType: this.filterType,
 					filter: this.checkedValues
 				})
 			},
@@ -181,6 +254,7 @@
 			handleSearchSubmit() {
 				this.closePopup()
 				this.$emit('change', {
+					filterType: this.filterType,
 					filter: this.filterValue
 				})
 			},
@@ -190,6 +264,21 @@
 				}
 				this.filterValue = ''
 				this.handleSearchSubmit()
+			},
+			handleRangeSubmit(isReset) {
+				this.closePopup()
+				this.$emit('change', {
+					filterType: this.filterType,
+					filter: isReset === true ? [] : [parseInt(this.gtValue), parseInt(this.ltValue)]
+				})
+			},
+			handleRangeReset() {
+				if (!this.canReset) {
+					return;
+				}
+				this.gtValue = ''
+				this.ltValue = ''
+				this.handleRangeSubmit(true)
 			}
 		}
 	}
@@ -260,6 +349,53 @@
 		background-color: #1890ff;
 	}
 
+	.icon-calendar {
+		color: #ddd;
+		width: 14px;
+		height: 16px;
+	}
+
+	.icon-calendar-0 {
+		height: 4px;
+		margin-top: 3px;
+		margin-bottom: 1px;
+		background-color: #ddd;
+		border-radius: 2px 2px 1px 1px;
+		position: relative;
+	}
+	.icon-calendar-0:before, .icon-calendar-0:after {
+		content: '';
+		position: absolute;
+		top: -3px;
+		width: 4px;
+		height: 3px;
+		border-radius: 1px;
+		background-color: #ddd;
+	}
+	.icon-calendar-0:before {
+		left: 2px;
+	}
+	.icon-calendar-0:after {
+		right: 2px;
+	}
+
+	.icon-calendar-1 {
+		height: 9px;
+		background-color: #ddd;
+		border-radius: 1px 1px 2px 2px;
+	}
+
+	.icon-calendar.active {
+		color: #1890ff;
+	}
+
+	.icon-calendar.active .icon-calendar-0,
+	.icon-calendar.active .icon-calendar-1,
+	.icon-calendar.active .icon-calendar-0:before,
+	.icon-calendar.active .icon-calendar-0:after {
+		background-color: #1890ff;
+	}
+
 	.uni-filter-dropdown {
 		position: relative;
 		font-weight: normal;
@@ -317,6 +453,22 @@
 	.search-input {
 		font-size: 12px;
 		border: 1px solid #f0f0f0;
+		border-radius: 3px;
+		padding: 2px 5px;
+		min-width: 150px;
+		text-align: left;
+	}
+
+	.input-label {
+		margin: 10px 10px 5px 10px;
+		text-align: left;
+	}
+
+	.input {
+		font-size: 12px;
+		border: 1px solid #f0f0f0;
+		border-radius: 3px;
+		margin: 10px;
 		padding: 2px 5px;
 		min-width: 150px;
 		text-align: left;
@@ -338,6 +490,7 @@
 	.btn-default {
 		border: 1px solid #ddd;
 	}
+
 	.btn-default.disable {
 		border-color: transparent;
 	}